+++ back_end/saolei/customranking/utils.py
@@ -1,2 +1,5 @@
+CUSTOM_PLUCK_CACHE_SIZE = 100
+
+
 def get_custom_pluck_cache_key(level: str) -> str:
     return f'customranking:pluck:{level}:top'
